feat(deps): lazy load dialog contents
