+++ src/components/Dialogs/DonationDialog/types.d.ts
@@ -1,0 +1,22 @@
+import {
+  ArticleDetailPublicQuery,
+  UserDonationRecipientFragment,
+} from '~/gql/graphql'
+
+export type Step =
+  | 'currencyChoice'
+  | 'walletSelect'
+  | 'setAmount'
+  | 'addCredit'
+  | 'complete'
+  | 'confirm'
+  | 'processing'
+  | 'resetPassword'
+  | 'setPaymentPassword'
+
+export interface BaseDonationDialogProps {
+  completeCallback?: () => void
+  recipient: UserDonationRecipientFragment
+  article: NonNullable<ArticleDetailPublicQuery['article']>
+  targetId: string
+}
