+++ src/components/Dialogs/ENSDialog/types.d.ts
@@ -1,0 +1,8 @@
+import { UserProfileUserPublicQuery } from '~/gql/graphql'
+
+export type Step = 'connectWallet' | 'walletSelect' | 'linkENS' | 'complete'
+
+export interface BaseENSDialogProps {
+  user: UserProfileUserPublicQuery['user']
+  completeCallback?: () => void
+}
