+++ src/components/Dialogs/LikeCoinDialog/Content/Binding/index.tsx
