+++ src/components/Dialogs/LikeCoinDialog/Content/Complete.tsx
