+++ src/components/Dialogs/LikeCoinDialog/Content/Generating/index.tsx
