+++ src/components/Dialogs/LikeCoinDialog/Content/Select/Hint.tsx
