+++ src/components/Dialogs/LikeCoinDialog/Content/Select/Intro.tsx
