+++ src/components/Dialogs/LikeCoinDialog/Content/Select/index.tsx
