+++ src/components/Dialogs/PayoutDialog/types.d.ts
@@ -1,0 +1,6 @@
+export type Step =
+  | 'complete'
+  | 'connectStripeAccount'
+  | 'confirm'
+  | 'processing'
+  | 'resetPassword'
