+++ src/components/Dialogs/SubscribeCircleDialog/index.tsx
@@ -1,4 +1,3 @@
-import gql from 'graphql-tag'
 import dynamic from 'next/dynamic'
 import { useContext, useEffect } from 'react'
 
@@ -7,66 +6,24 @@ import { analytics } from '~/common/utils'
 import {
   Dialog,
   Spinner,
-  Translate,
   useDialogSwitch,
   useEventListener,
   useStep,
   ViewerContext,
 } from '~/components'
-import { CircleDigest } from '~/components/CircleDigest'
-import {
-  DigestRichCirclePrivateFragment,
-  DigestRichCirclePublicFragment,
-} from '~/gql/graphql'
 
-import Complete from './Complete'
+import { fragments } from './gql'
+import { BaseSubscribeCircleDialogProps, Step } from './types'
 
-type Step =
-  | 'setPaymentPassword'
-  | 'subscribeCircle'
-  | 'resetPassword'
-  | 'complete'
-
-interface SubscribeCircleDialogProps {
-  circle: DigestRichCirclePublicFragment & DigestRichCirclePrivateFragment
+type SubscribeCircleDialogProps = BaseSubscribeCircleDialogProps & {
   children?: ({ openDialog }: { openDialog: () => void }) => React.ReactNode
 }
 
-const DynamicPaymentResetPasswordForm = dynamic(
-  () => import('~/components/Forms/PaymentForm/ResetPassword'),
-  { loading: Spinner }
-)
-const DynamicPaymentSetPasswordForm = dynamic(
-  () => import('~/components/Forms/PaymentForm/SetPassword'),
-  { loading: Spinner }
-)
-const DynamicSubscribeCircleForm = dynamic(
-  () => import('~/components/Forms/PaymentForm/SubscribeCircle'),
-  { loading: Spinner }
-)
-
-const fragments = {
-  circle: {
-    public: gql`
-      fragment SubscribeCirclePublic on Circle {
-        id
-        ...DigestRichCirclePublic
-      }
-      ${CircleDigest.Rich.fragments.circle.public}
-    `,
-    private: gql`
-      fragment SubscribeCirclePrivate on Circle {
-        id
-        ...DigestRichCirclePrivate
-      }
-      ${CircleDigest.Rich.fragments.circle.private}
-    `,
-  },
-}
+const DynamicContent = dynamic(() => import('./Content'), { loading: Spinner })
 
 const BaseSubscribeCircleDialog = ({
-  circle,
   children,
+  ...restProps
 }: SubscribeCircleDialogProps) => {
   const viewer = useContext(ViewerContext)
   const {
@@ -85,17 +42,6 @@ const BaseSubscribeCircleDialog = ({
     baseOpenDialog()
   }
 
-  const ContinueSubscribeButton = (
-    <Dialog.Footer.Button onClick={() => forward('subscribeCircle')}>
-      <Translate zh_hant="回到訂閱" zh_hans="回到订阅" />
-    </Dialog.Footer.Button>
-  )
-
-  const isSetPaymentPassword = currStep === 'setPaymentPassword'
-  const isSubscribeCircle = currStep === 'subscribeCircle'
-  const isComplete = currStep === 'complete'
-  const isResetPassword = currStep === 'resetPassword'
-
   useEffect(() => {
     analytics.trackEvent('view_subscribe_circle_dialog', { step: currStep })
   }, [currStep])
@@ -107,52 +53,14 @@ const BaseSubscribeCircleDialog = ({
       {children && children({ openDialog })}
 
       <Dialog size="sm" isOpen={show} onDismiss={closeDialog}>
-        <Dialog.Header
-          leftButton={
-            prevStep ? <Dialog.Header.BackButton onClick={back} /> : <span />
-          }
-          rightButton={
-            <Dialog.Header.CloseButton
-              closeDialog={closeDialog}
-              textId="close"
-            />
-          }
-          title={
-            isSetPaymentPassword
-              ? 'paymentPassword'
-              : isComplete
-              ? 'successSubscribeCircle'
-              : isResetPassword
-              ? 'resetPaymentPassword'
-              : 'subscribeCircle'
-          }
+        <DynamicContent
           closeDialog={closeDialog}
-          closeTextId="close"
-          mode={isComplete ? 'inner' : undefined}
+          forward={forward}
+          back={back}
+          currStep={currStep}
+          prevStep={prevStep}
+          {...restProps}
         />
-
-        {isSetPaymentPassword && (
-          <DynamicPaymentSetPasswordForm
-            submitCallback={() => forward('subscribeCircle')}
-          />
-        )}
-
-        {isSubscribeCircle && (
-          <DynamicSubscribeCircleForm
-            circle={circle}
-            submitCallback={() => forward('complete')}
-            switchToResetPassword={() => forward('resetPassword')}
-          />
-        )}
-
-        {isResetPassword && (
-          <DynamicPaymentResetPasswordForm
-            callbackButtons={ContinueSubscribeButton}
-            closeDialog={closeDialog}
-          />
-        )}
-
-        {isComplete && <Complete circle={circle} />}
       </Dialog>
     </>
   )
