+++ src/components/Dialogs/SubscribeCircleDialog/types.d.ts
@@ -1,0 +1,9 @@
+export type Step =
+  | 'setPaymentPassword'
+  | 'subscribeCircle'
+  | 'resetPassword'
+  | 'complete'
+
+export interface BaseSubscribeCircleDialogProps {
+  circle: DigestRichCirclePublicFragment & DigestRichCirclePrivateFragment
+}
