+++ src/views/Circle/Profile/styles.css
@@ -1,4 +1,6 @@
 .buttons {
+  display: inline-flex;
+
   & :global(> * + *) {
     margin-left: var(--spacing-base);
   }
